# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gold Mailer API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthMessage {
  message: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface VerifyEmailInput {
  email: string;
  code: string;
}

export interface EmailInput {
  email: string;
}

export interface ResetPasswordInput {
  email: string;
  code: string;
  /** @minLength 6 */
  newPassword: string;
}

export interface ChangePasswordInput {
  currentPassword: string;
  /** @minLength 6 */
  newPassword: string;
}

export interface ChangeEmailInput {
  newEmail: string;
}

export interface User {
  id: number;
  email: string;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  middleName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  isVerified: boolean;
  profileComplete: boolean;
  cardAdded: boolean;
  balance: number;
  createdAt: string;
}

export interface LoginResult {
  user: User;
  message: string;
}

export interface ProfileInput {
  firstName: string;
  middleName?: string;
  lastName: string;
  age?: number;
  gender?: string;
  avatarUrl?: string;
}

export interface DashboardData {
  balance: number;
  totalStaked: number;
  totalProfit: number;
  activeStakes: number;
  dailyRewardAvailable: boolean;
  pendingDeposits: number;
  pendingWithdrawals: number;
}

export interface Card {
  id: number;
  cardholderName: string;
  lastFour: string;
  expiryDate: string;
  /** @nullable */
  billingCity?: string | null;
  /** @nullable */
  billingCountry?: string | null;
  createdAt: string;
}

export interface CardFull {
  id: number;
  cardholderName: string;
  cardNumber: string;
  expiryDate: string;
  cvv: string;
  billingAddress1: string;
  /** @nullable */
  billingAddress2?: string | null;
  billingCity: string;
  billingState: string;
  billingCountry: string;
  billingZip: string;
  /** @nullable */
  aptNumber?: string | null;
}

export interface CardInput {
  cardholderName: string;
  cardNumber: string;
  expiryDate: string;
  cvv: string;
  billingAddress1: string;
  billingAddress2?: string;
  billingCity: string;
  billingState: string;
  billingCountry: string;
  billingZip: string;
  aptNumber?: string;
}

export interface Stake {
  id: number;
  amount: number;
  profit: number;
  status: string;
  startDate: string;
  endDate: string;
  dailyClaimedToday: boolean;
  totalDailyClaimed: number;
  daysRemaining?: number;
}

export interface StakeInput {
  amount: number;
}

export interface RewardClaim {
  message: string;
  amount: number;
  newBalance: number;
}

export interface Transaction {
  id: number;
  type: string;
  amount: number;
  status: string;
  /** @nullable */
  transactionId?: string | null;
  /** @nullable */
  bankName?: string | null;
  /** @nullable */
  accountNumber?: string | null;
  /** @nullable */
  accountName?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface DepositInput {
  amount: number;
  transactionId: string;
}

export interface WithdrawalInput {
  amount: number;
  bankName: string;
  accountNumber: string;
  accountName: string;
}

export interface DepositAccount {
  bankName: string;
  accountNumber: string;
  accountName: string;
}

export interface AdminUser {
  id: number;
  email: string;
  plainPassword: string;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  balance: number;
  isVerified: boolean;
  profileComplete: boolean;
  createdAt: string;
}

export interface BalanceTopUp {
  amount: number;
}