# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Gold Mailer API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered, verification email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMessage"
        "400":
          description: Invalid input or email already exists
  /auth/verify-email:
    post:
      operationId: verifyEmail
      tags: [auth]
      summary: Verify email with OTP code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailInput"
      responses:
        "200":
          description: Email verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMessage"
        "400":
          description: Invalid or expired code
  /auth/resend-verification:
    post:
      operationId: resendVerification
      tags: [auth]
      summary: Resend verification code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailInput"
      responses:
        "200":
          description: Code resent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMessage"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          description: Invalid credentials
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailInput"
      responses:
        "200":
          description: Reset email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMessage"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with OTP code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMessage"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
...
  /cards:
    get:
      operationId: getCards
      tags: [card]
      summary: Get all user virtual cards
      responses:
        "200":
          description: List of cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Card"

    post:
      operationId: addCard
      tags: [card]
      summary: Add a new virtual card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardInput"
      responses:
        "201":
          description: Card added with ₦3000 signup bonus
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
  /cards/{id}:
    get:
      operationId: getCard
      tags: [card]
      summary: Get full card details including sensitive info
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Full card details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardFull"
...
  /admin/users/{id}:
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Delete a user account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMessage"
  /admin/users/{id}/balance:
    post:
      operationId: adminTopUpBalance
      tags: [admin]
      summary: Top up user balance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BalanceTopUp"
      responses:
        "200":
          description: Balance updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
...
  /admin/transactions/{id}/approve:
    post:
      operationId: adminApproveTransaction
      tags: [admin]
      summary: Approve a deposit or withdrawal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /admin/transactions/{id}/decline:
    post:
      operationId: adminDeclineTransaction
      tags: [admin]
      summary: Decline a deposit or withdrawal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction declined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /admin/deposit-account:
    put:
      operationId: adminSetDepositAccount
      tags: [admin]
      summary: Set the deposit bank account details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositAccount"
      responses:
        "200":
          description: Deposit account updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositAccount"
components:
  schemas:
...
    AuthMessage:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    RegisterInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    VerifyEmailInput:
      type: object
      required: [email, code]
      properties:
        email:
          type: string
          format: email
        code:
          type: string
    EmailInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ResetPasswordInput:
      type: object
      required: [email, code, newPassword]
      properties:
        email:
          type: string
        code:
          type: string
        newPassword:
          type: string
          minLength: 6
...
    LoginResult:
      type: object
      required: [user, message]
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
    User:
      type: object
      required: [id, email, isVerified, profileComplete, cardAdded, balance, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: ["string", "null"]
        middleName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        age:
          type: ["integer", "null"]
        gender:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isVerified:
          type: boolean
        profileComplete:
          type: boolean
        cardAdded:
          type: boolean
        balance:
          type: number
        createdAt:
          type: string
...
    Card:
      type: object
      required: [id, cardholderName, lastFour, expiryDate, createdAt]
      properties:
        id:
          type: integer
        cardholderName:
          type: string
        lastFour:
          type: string
        expiryDate:
          type: string
        billingCity:
          type: ["string", "null"]
        billingCountry:
          type: ["string", "null"]
        createdAt:
          type: string
    CardFull:
      type: object
      required: [id, cardholderName, cardNumber, expiryDate, cvv, billingAddress1, billingCity, billingState, billingCountry, billingZip]
      properties:
        id:
          type: integer
        cardholderName:
          type: string
        cardNumber:
          type: string
        expiryDate:
          type: string
        cvv:
          type: string
        billingAddress1:
          type: string
        billingAddress2:
          type: ["string", "null"]
        billingCity:
          type: string
        billingState:
          type: string
        billingCountry:
          type: string
        billingZip:
          type: string
        aptNumber:
          type: ["string", "null"]
    CardInput:
      type: object
      required: [cardholderName, cardNumber, expiryDate, cvv, billingAddress1, billingCity, billingState, billingCountry, billingZip]
      properties:
        cardholderName:
          type: string
        cardNumber:
          type: string
        expiryDate:
          type: string
        cvv:
          type: string
        billingAddress1:
          type: string
        billingAddress2:
          type: string
        billingCity:
          type: string
        billingState:
          type: string
        billingCountry:
          type: string
        billingZip:
          type: string
        aptNumber:
          type: string
...
    Transaction:
      type: object
      required: [id, type, amount, status, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        amount:
          type: number
        status:
          type: string
        transactionId:
          type: ["string", "null"]
        bankName:
          type: ["string", "null"]
        accountNumber:
          type: ["string", "null"]
        accountName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    DepositAccount:
      type: object
      required: [bankName, accountNumber, accountName]
      properties:
        bankName:
          type: string
        accountNumber:
          type: string
        accountName:
          type: string
    AdminUser:
      type: object
      required: [id, email, plainPassword, balance, isVerified, profileComplete, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        plainPassword:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        balance:
          type: number
        isVerified:
          type: boolean
        profileComplete:
          type: boolean
        createdAt:
          type: string
    BalanceTopUp:
      type: object
      required: [amount]
      properties:
        amount:
          type: number