# lib/api-zod/src/generated/types/card.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gold Mailer API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Card {
  id: number;
  cardholderName: string;
  lastFour: string;
  expiryDate: string;
  /** @nullable */
  billingCity?: string | null;
  /** @nullable */
  billingCountry?: string | null;
  createdAt: string;
}