# lib/api-zod/src/generated/types/depositAccount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gold Mailer API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DepositAccount {
  bankName: string;
  accountNumber: string;
  accountName: string;
}