# lib/api-zod/src/generated/types/loginResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gold Mailer API specification
 * OpenAPI spec version: 0.1.0
 */
import type { User } from "./user";

export interface LoginResult {
  user: User;
  message: string;
}