# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gold Mailer API specification
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  email: string;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  middleName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  isVerified: boolean;
  profileComplete: boolean;
  cardAdded: boolean;
  balance: number;
  createdAt: string;
}