# lib/api-zod/src/generated/types/verifyEmailInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gold Mailer API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VerifyEmailInput {
  email: string;
  code: string;
}